♻️ Del ut i flere filer
